+++ core/src/gravity.rs
@@ -1,4 +1,4 @@
-use bevy_math::{Vec2, Vec3};
+use bevy::math::{Vec2, Vec3};
 
 /// Resource that defines world's gravity.
 ///