+++ rapier/src/lib.rs
@@ -13,11 +13,10 @@ pub extern crate rapier2d as rapier;
 #[cfg(feature = "3d")]
 pub extern crate rapier3d as rapier;
 
-use bevy_app::{AppBuilder, Plugin};
-use bevy_core::FixedTimestep;
-use bevy_ecs::{IntoChainSystem, IntoSystem, Schedule, SystemStage};
+use bevy::app::{AppBuilder, Plugin};
+use bevy::ecs::{IntoSystem, Schedule, SystemStage};
 
-use heron_core::{CollisionEvent, Gravity};
+use heron_core::CollisionEvent;
 
 use crate::body::HandleMap;
 use crate::rapier::dynamics::{IntegrationParameters, JointSet, RigidBodyHandle, RigidBodySet};
@@ -33,11 +32,9 @@ mod velocity;
 
 #[allow(unused)]
 mod stage {
-    pub(crate) const START: &str = "heron-start";
     pub(crate) const PRE_STEP: &str = "heron-pre-step";
     pub(crate) const STEP: &str = "heron-step";
     pub(crate) const POST_STEP: &str = "heron-post-step";
-    pub(crate) const DEBUG: &str = "heron-debug";
 }
 
 /// Plugin to install in order to enable collision detection and physics behavior, powered by rapier.
@@ -54,7 +51,7 @@ pub struct RapierPlugin {
     pub parameters: IntegrationParameters,
 }
 
-/// Components automatically register, by the plugin that references the body in rapier's world
+/// Components automatically register, by the plugin that references the body in rapier's world.
 ///
 /// It can be used to get direct access to rapier's world.
 #[derive(Debug, Copy, Clone)]
@@ -64,11 +61,11 @@ pub struct BodyHandle {
 }
 
 impl RapierPlugin {
-    /// Configure how many times per second the physics world needs to be updated
+    /// Configure how many times per second the physics world needs to be updated.
     ///
     /// # Panics
     ///
-    /// Panic if the number of `steps_per_second` is 0
+    /// Panic if the number of `steps_per_second` is 0.
     pub fn from_steps_per_second(steps_per_second: u8) -> Self {
         assert!(
             steps_per_second > 0,
@@ -105,59 +102,56 @@ impl From<IntegrationParameters> for RapierPlugin {
 
 impl Plugin for RapierPlugin {
     fn build(&self, app: &mut AppBuilder) {
-        app.resources_mut().get_or_insert_with(Gravity::default);
-
-        app.init_resource::<PhysicsPipeline>()
-            .init_resource::<HandleMap>()
-            .add_event::<CollisionEvent>()
-            .add_resource(self.parameters)
-            .add_resource(BroadPhase::new())
-            .add_resource(NarrowPhase::new())
-            .add_resource(RigidBodySet::new())
-            .add_resource(ColliderSet::new())
-            .add_resource(JointSet::new())
-            .add_stage_after(
-                bevy_app::stage::POST_UPDATE,
-                stage::PRE_STEP,
-                SystemStage::serial()
-                    .with_system(body::remove.system())
-                    .with_system(body::recreate_collider.system())
-                    .with_system(body::update_rapier_position.system())
-                    .with_system(body::update_rapier_status.system())
-                    .with_system(velocity::update_rapier_velocity.system())
-                    .with_system(acceleration::update_rapier_force_and_torque.system())
-                    .with_system(body::create.system())
-            )
-            .add_stage_after(stage::PRE_STEP, "heron-step-and-post-step", {
-                let mut schedule = Schedule::default();
-
-                if let Some(steps_per_second) = self.step_per_second {
-                    schedule =
-                        schedule.with_run_criteria(FixedTimestep::steps_per_second(steps_per_second))
-                }
-
-                schedule.with_stage(
-                    stage::STEP,
+        app.add_plugin(heron_core::CorePlugin {
+            steps_per_second: self.step_per_second,
+        })
+        .init_resource::<PhysicsPipeline>()
+        .init_resource::<HandleMap>()
+        .add_event::<CollisionEvent>()
+        .add_resource(self.parameters)
+        .add_resource(BroadPhase::new())
+        .add_resource(NarrowPhase::new())
+        .add_resource(RigidBodySet::new())
+        .add_resource(ColliderSet::new())
+        .add_resource(JointSet::new())
+        .stage(heron_core::stage::ROOT, |schedule: &mut Schedule| {
+            schedule
+                .add_stage(
+                    "heron-remove-invalid-bodies",
+                    SystemStage::serial().with_system(body::remove_bodies.system()),
+                )
+                .add_stage(
+                    "heron-pre-step",
                     SystemStage::serial()
+                        .with_system(
+                            bevy::transform::transform_propagate_system::transform_propagate_system
+                                .system(),
+                        )
+                        .with_system(body::remove.system())
+                        .with_system(body::update_rapier_position.system())
+                        .with_system(velocity::update_rapier_velocity.system())
+                        .with_system(body::update_rapier_status.system())
+                        .with_system(acceleration::update_rapier_force_and_torque.system())
+                        .with_system(body::create.system()),
+                )
+                .add_stage(
+                    "heron-step",
+                    SystemStage::serial()
+                        .with_system(velocity::apply_velocity_to_kinematic_bodies.system())
                         .with_system(pipeline::step.system()),
                 )
-                    .with_stage(
-                        stage::POST_STEP,
-                        SystemStage::parallel()
-                            .with_system(
-                                body::update_bevy_transform.system().chain(
-                                    bevy_transform::transform_propagate_system::transform_propagate_system
-                                        .system()
-                                )
-                            )
-                            .with_system(velocity::update_velocity_component.system())
-                    )
-            });
+                .add_stage(
+                    "heron-post-step",
+                    SystemStage::parallel()
+                        .with_system(body::update_bevy_transform.system())
+                        .with_system(velocity::update_velocity_component.system()),
+                )
+        });
     }
 }
 
 impl BodyHandle {
-    /// Creates the new `BodyHandle`
+    /// Creates the new `BodyHandle`.
     #[must_use]
     pub fn new(rigid_body: RigidBodyHandle, collider: ColliderHandle) -> BodyHandle {
         BodyHandle {
@@ -166,13 +160,13 @@ impl BodyHandle {
         }
     }
 
-    /// Returns the rapier's rigid body handle
+    /// Returns the rapier's rigid body handle.
     #[must_use]
     pub fn rigid_body(&self) -> RigidBodyHandle {
         self.rigid_body
     }
 
-    /// Returns the rapier's collider handle
+    /// Returns the rapier's collider handle.
     #[must_use]
     pub fn collider(&self) -> ColliderHandle {
         self.collider